[fix] #15 - 오류 수정
