+++ POTI-iOS/POTI-iOS/Resource/Info.plist
@@ -4,9 +4,9 @@
 <dict>
 	<key>UIAppFonts</key>
 	<array>
-		<string>Pretendard-Bold</string>
-		<string>Pretendard-Medium</string>
-		<string>Pretendard-SemiBold</string>
+		<string>Pretendard-Bold.otf</string>
+		<string>Pretendard-Medium.otf</string>
+		<string>Pretendard-SemiBold.otf</string>
 	</array>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
